# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SocialApp — unified social network + messaging platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /groups:
    get:
      tags: [Groups]
      operationId: getMyGroups
      summary: Get groups I belong to
      responses:
        '200':
          description: Groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
    post:
      tags: [Groups]
      operationId: createGroup
      summary: Create a group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupInput'
      responses:
        '201':
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
  /groups/{groupId}:
    get:
      tags: [Groups]
      operationId: getGroup
      summary: Get group details
      parameters:
        - name: groupId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
    patch:
      tags: [Groups]
      operationId: updateGroup
      summary: Update group details
      parameters:
        - name: groupId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupUpdate'
      responses:
        '200':
          description: Updated group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
  /groups/{groupId}/members:
    post:
      tags: [Groups]
      operationId: addGroupMember
      summary: Add a member to a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberInput'
      responses:
        '200':
          description: Member added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
  /groups/{groupId}/members/{userId}:
    delete:
      tags: [Groups]
      operationId: removeGroupMember
      summary: Remove a member from a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema: { type: string }
        - name: userId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Member removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
  /groups/{groupId}/messages:
    get:
      tags: [Groups]
      operationId: getGroupMessages
      summary: Get messages in a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema: { type: string }
        - name: cursor
          in: query
          required: false
          schema: { type: string }
      responses:
        '200':
          description: Messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageListPage'
    post:
      tags: [Groups]
      operationId: sendGroupMessage
      summary: Send a message to a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageInput'
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'

  # ── Channels ──────────────────────────────────────────────────────────────
...
components:
...
  schemas:
...
    UserSummary:
      type: object
      required: [id, username, displayName]
      properties:
        id: { type: string }
        username: { type: string }
        displayName: { type: string }
        avatarUrl: { type: ["string", "null"] }
        isFollowing: { type: boolean }
...
    Message:
      type: object
      required: [id, senderId, content, isRead, createdAt]
      properties:
        id: { type: string }
        senderId: { type: string }
        sender:
          oneOf:
            - $ref: '#/components/schemas/UserSummary'
            - type: 'null'
        content: { type: string }
        mediaUrl: { type: ["string", "null"] }
        mediaType: { type: ["string", "null"], enum: [image, video, audio, file, "null"] }
        isRead: { type: boolean }
        createdAt: { type: string, format: date-time }
    MessageInput:
      type: object
      required: [content]
      properties:
        content: { type: string, minLength: 1 }
        mediaUrl: { type: string }
        mediaType: { type: string, enum: [image, video, audio, file] }
    MessageListPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        nextCursor: { type: ["string", "null"] }

    # ── Groups ────────────────────────────────────────────────────────────────
    Group:
      type: object
      required: [id, name, membersCount, isAdmin, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        membersCount: { type: integer }
        members:
          type: array
          items:
            $ref: '#/components/schemas/UserSummary'
        isAdmin: { type: boolean }
        lastMessage:
          oneOf:
            - $ref: '#/components/schemas/Message'
            - type: 'null'
        unreadCount: { type: integer }
        createdAt: { type: string, format: date-time }
    GroupInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        avatarUrl: { type: string }
        memberIds:
          type: array
          items: { type: string }
    GroupUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        avatarUrl: { type: string }
    MemberInput:
      type: object
      required: [userId]
      properties:
        userId: { type: string, minLength: 1 }

    # ── Channels ──────────────────────────────────────────────────────────────
...